% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.create_register_pmc_title}
\alias{.create_register_pmc_title}
\title{Identify registration titles - specific with no negation}
\usage{
.create_register_pmc_title()
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Trial registration: ..."
}
