% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{.create_synonyms}
\alias{.create_synonyms}
\title{A list of word synonyms}
\usage{
.create_synonyms()
}
\value{
A list of synonyms to words of interest
}
\description{
Contains the synonyms to words being used throughout the package.
}
