% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.encase}
\alias{.encase}
\title{Encace words within parentheses and OR statements}
\usage{
.encase(x)
}
\arguments{
\item{x}{A vector of strings.}
}
\value{
A string of words separate by OR statements within parentheses.
}
\description{
Returns a string of words separated by OR statements within parentheses.
}
