% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.first_capital}
\alias{.first_capital}
\title{Create a regular expression where the first letter is capital}
\usage{
.first_capital(x, location = "both")
}
\arguments{
\item{x}{A vector of strings.}

\item{location}{Whether to "start", "end" or "both" with a capital letter.}
}
\value{
A string pattern.
}
\description{
Returns a regular expression that necessitates that the first letter is
    capital and the rest can be any case.
}
