% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.get_coi_pmc_fn}
\alias{.get_coi_pmc_fn}
\title{Identify elements with COI attributes}
\usage{
.get_coi_pmc_fn(article_xml)
}
\arguments{
\item{article_xml}{The text as an xml_document.}
}
\value{
The title and its related text as a string.
}
\description{
Extract elements with COI attributes from NLM XLMs.
}
