% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.get_coi_pmc_title}
\alias{.get_coi_pmc_title}
\title{Identify COI titles using XML labels}
\usage{
.get_coi_pmc_title(article_xml, dict)
}
\arguments{
\item{article_xml}{The text as an xml_document.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
The title and its related text as a string.
}
\description{
Extract XML titles related to COI statements and all text children.
}
