% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund_pmc.R
\name{.get_fund_pmc_group}
\alias{.get_fund_pmc_group}
\title{Identify funding group elements in NLM XML files}
\usage{
.get_fund_pmc_group(article_xml)
}
\arguments{
\item{article_xml}{An NLM XML as an xml_document.}
}
\value{
The text of interest as a list indicating whether it was found, the
    quoated institutes and the actual statement as a string.
}
\description{
Identify and return the text found in funding group elements of NLM XML
    files. Articles with funding-group elements also have a Funding
    indication on PubMed.
}
