% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund_pmc.R
\name{.get_fund_pmc_source}
\alias{.get_fund_pmc_source}
\title{Identify funding source elements in NLM XML files}
\usage{
.get_fund_pmc_source(article_xml)
}
\arguments{
\item{article_xml}{An NLM XML as an xml_document.}
}
\value{
The text of interest as a string.
}
\description{
Identify and return the text found in funding source elements of NLM XML
    files. Some funding sources are not included in the funding-group
    elements - these are the sources that this function attempts to capture.
    It seems that all articles with such elements also have a Funding
    statement on PubMed.
}
