% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund_pmc.R
\name{.get_fund_pmc_title}
\alias{.get_fund_pmc_title}
\title{Identify funding titles using XML labels}
\usage{
.get_fund_pmc_title(article_xml)
}
\arguments{
\item{article_xml}{The text as an xml_document.}
}
\value{
The title and its related text as a string.
}
\description{
Extract XML titles related to funding and all text children.
}
