% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{.get_text}
\alias{.get_text}
\title{Get the desired text from the xml_document}
\usage{
.get_text(article_xml, xpath, find_first)
}
\arguments{
\item{article_xml}{An xml_document from `xml2::read_xml`.}

\item{xpath}{The XPath as a character, e.g. "id_info/nct_id".}

\item{find_first}{TRUE to find first mention, FALSE to find all mentions.}
}
\value{
The desired text as a character; if not found, then `character()`
}
\description{
Returns the text desired according to xpath.
}
