% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{.get_xml}
\alias{.get_xml}
\title{Read an XML file into an xml_document}
\usage{
.get_xml(filename, remove_ns = F)
}
\arguments{
\item{filename}{The filepath to the PMC XML file of interest.}

\item{remove_ns}{Whether to remove the XML namespace or not (default = F).}
}
\value{
The PMC XML as an xml_document.
}
\description{
Returns the file as an xml_document.
}
