% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.negate_coi_1}
\alias{.negate_coi_1}
\title{Negate statements that mention conflict but are not COI}
\usage{
.negate_coi_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
A boolean indicating whether a disclosure should be retained.
}
\description{
Keep statements such as "No conficts of interest reported", but do not
     keep statements such as "The conflicts of interest are becoming
     commoner."
}
