% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.negate_disclosure_1}
\alias{.negate_disclosure_1}
\title{Negate disclosures that are funding statements}
\usage{
.negate_disclosure_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
A boolean indicating whether a disclosure should be retained.
}
\description{
Keep statements such as "No financial disclosures reported.", but do not
     keep statements such as "The authors disclose not funding received."
}
