% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.negate_disclosure_2}
\alias{.negate_disclosure_2}
\title{Negate disclosures that are neither COI nor funding statements}
\usage{
.negate_disclosure_2(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
A boolean indicating whether a disclosure should be retained.
}
\description{
Remove mentions of disclosures such as "Patient Information Disclosure".
}
