% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.negate_reg_title_2}
\alias{.negate_reg_title_2}
\title{Negate titles that mention that there was no registration}
\usage{
.negate_reg_title_2(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Negate mentions such as "Clinical Trial Registration: N/A"
}
