% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.negate_registry_1}
\alias{.negate_registry_1}
\title{Negate unwanted mentions of registry}
\usage{
.negate_registry_1(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Negate mentions of registry such as: "Ethics approval and consent to
    participate This study was approved by the Institutional Review Board of
    Chang Gung Memorial Hospital under registry number 201601023B0."
}
