% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_replication.R
\name{.negate_replication_1}
\alias{.negate_replication_1}
\title{Remove negated replication mentions}
\usage{
.negate_replication_1(article)
}
\arguments{
\item{article}{A character vector of matching paragraphs.}
}
\value{
Logical vector; TRUE where the match is a negation (to be excluded).
}
\description{
Removes mentions such as "failed to replicate" or "could not replicate".
}
