% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.obliterate_apostrophe_1}
\alias{.obliterate_apostrophe_1}
\title{Remove apostrophe}
\usage{
.obliterate_apostrophe_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Removes commas to make ease creation of regular expressions. After
    implmenting this function, "ball's" should become balls and l'Alba
    should become lAlba and balls' into balls.
}
