% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.obliterate_contribs}
\alias{.obliterate_contribs}
\title{Remove author contribution statements}
\usage{
.obliterate_contribs(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Returns the text without the author contribution statements.
}
