% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.obliterate_hash_1}
\alias{.obliterate_hash_1}
\title{Remove hash}
\usage{
.obliterate_hash_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Removes hashes to make ease creation of regular expressions.
}
