% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.obliterate_honoraria_1}
\alias{.obliterate_honoraria_1}
\title{Remove irrelevat mentions of honoraria}
\usage{
.obliterate_honoraria_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
A boolean indicating whether a disclosure should be retained.
}
\description{
Remove mentions of honoraria such as "Patients received honoraria".
}
