% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.obliterate_line_break_1}
\alias{.obliterate_line_break_1}
\title{Remove break of line tags.}
\usage{
.obliterate_line_break_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs with no remaining line break tags.
}
\description{
Removes missed breaks of line
}
