% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.obliterate_punct_1}
\alias{.obliterate_punct_1}
\title{Remove uninformative punctuation}
\usage{
.obliterate_punct_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without backlashes.
}
\description{
Removes irrelevant puncutation to ease creation of regular expressions.
}
