% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.obliterate_references_1}
\alias{.obliterate_references_1}
\title{Remove references}
\usage{
.obliterate_references_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Removes mentions such as: "An active observational cohort study was
    conducted as previously reported (ClinicalTrials.gov identifier
    NCT01280162) [16]."
}
