% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.obliterate_refs_2}
\alias{.obliterate_refs_2}
\title{Remove mentions of previously reported registered studies}
\usage{
.obliterate_refs_2(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Removes mentions such as: "An active o <- servational cohort study was
    conducted as previously reported (ClinicalTrials.gov identifier
    NCT01280162) [16]."
}
