% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{.reroot_xml}
\alias{.reroot_xml}
\title{Reconfigure the PMC XML so that the top node is "article".}
\usage{
.reroot_xml(article_xml)
}
\arguments{
\item{article_xml}{The article as an xml_document.}
}
\value{
The xml_document reconfigure to start with the "article" node.
}
\description{
PMC uses different XML top nodes, depending on what service is used to
    download an article. This function standardizes the XML files so that
    they all start from the same root.
}
