% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund_pmc.R
\name{.rt_fund_pmc}
\alias{.rt_fund_pmc}
\title{Identify and extract Funding statements in PMC XML files.}
\usage{
.rt_fund_pmc(article_ls, pmc_fund_ls)
}
\arguments{
\item{article_ls}{A PMC XML as a list of strings.}

\item{pmc_fund_ls}{A list of results from the `.get_fund_pmc` function.}
}
\value{
A dataframe of results.
}
\description{
Takes a PMC XML file as a list of strings and returns data related to the
    presence of a Funding statement, including whether a Funding statement
    exists. If a Funding statement exists, it extracts it. This is a modified
    version of the `rt_fund_pmc` designed for integration with `rt_all_pmc`.
}
