% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_novelty_pmc.R
\name{.rt_novelty_pmc}
\alias{.rt_novelty_pmc}
\title{Identify novelty claims in PMC XML files.}
\usage{
.rt_novelty_pmc(article_ls)
}
\arguments{
\item{article_ls}{A PMC XML as a list of strings (from \code{.get_article_txt}).}
}
\value{
A named list of results.
}
\description{
Takes a PMC XML file as a list of article sections and returns data related
    to the presence of novelty claims. This is the internal version designed
    for integration with \code{rt_all_pmc}.
}
