% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.rt_register_pmc}
\alias{.rt_register_pmc}
\title{Identify and extract Registration statements in PMC XML files.}
\usage{
.rt_register_pmc(article_ls, pmc_reg_ls, dict)
}
\arguments{
\item{article_ls}{A PMC XML as a list of strings.}

\item{pmc_reg_ls}{A list of results from the `.get_register_pmc` function.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
A dataframe of results.
}
\description{
Takes a PMC XML file as a list of strings and returns data related to the
    presence of a Registration statement, including whether such a statement
    exists. If a Registration statement exists, it extracts it. This is a
    modified version of the `rt_register_pmc` designed for integration with
    `rt_all_pmc`.
}
