% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_replication_pmc.R
\name{.rt_replication_pmc}
\alias{.rt_replication_pmc}
\title{Identify replication components in PMC XML files.}
\usage{
.rt_replication_pmc(article_ls)
}
\arguments{
\item{article_ls}{A PMC XML as a list of strings (from \code{.get_article_txt}).}
}
\value{
A named list of results.
}
\description{
Takes a PMC XML file as a list of article sections and returns data related
    to the presence of a replication or validation component. This is the
    internal version designed for integration with \code{rt_all_pmc}.
}
