% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.title_strict}
\alias{.title_strict}
\title{Create a regex for titles}
\usage{
.title_strict(x, within_text = F)
}
\arguments{
\item{x}{A vector of strings.}

\item{within_text}{Boolean defines whether a regex typical to a title found      within text should be created or not.}
}
\value{
A vector of strings with a suffix attached.
}
\description{
Returns words designed to identify titles.
}
