% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_data_code_pmc.R
\name{.tokenize}
\alias{.tokenize}
\title{Tokenize NLM XML files}
\usage{
.tokenize(article)
}
\arguments{
\item{article}{The article as a string}
}
\value{
List with one element per document. Each document is split into its
    sentences and saved as a vector of strings.
}
\description{
The code for this was modified from the oddpub package.
}
