% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.where_methods_txt}
\alias{.where_methods_txt}
\title{Find the Methods section}
\usage{
.where_methods_txt(article)
}
\arguments{
\item{article}{The text as a vector of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Find the index of the start of the Methods section.
}
