% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.where_refs_txt}
\alias{.where_refs_txt}
\title{Find the index of the references}
\usage{
.where_refs_txt(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the start and finish of this section.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
