% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_benefit_1}
\alias{.which_benefit_1}
\title{Identify mentions of receiving benefits}
\usage{
.which_benefit_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of elements with phrase of interest
}
\description{
Extract mentions of benefits, e.g. "SS has received benefits of commercial
    nature from GSK."
}
