% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_brief}
\alias{.which_brief}
\title{Identify mentions of brief explanation}
\usage{
.which_brief(article)
}
\arguments{
\item{article}{The text as a vector of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract mentions of COI statements that mention "Brief explanation."
}
