% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_coi_1}
\alias{.which_coi_1}
\title{Identify standard mentions of COI statements}
\usage{
.which_coi_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract mentions of COI statements that contain standard phrases, such as
    "Conflicts of Interest".
}
