% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_coi_2}
\alias{.which_coi_2}
\title{Identify less explicit mentions of COI statements.}
\usage{
.which_coi_2(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of elements with phrase of interest
}
\description{
Extract mentions of COI statements that are less standard, e.g. "We declare
    that we have no financial or other conflicts of interest."
}
