% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_commercial_1}
\alias{.which_commercial_1}
\title{Identify less explicit mentions of commercial interest}
\usage{
.which_commercial_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of elements with phrase of interest
}
\description{
Extract mentions of COI statements that are less standard and mention
    commercial interests, e.g. "The authors of this study declare a financial
    relationship with GSK."
}
