% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_ct_2}
\alias{.which_ct_2}
\title{Identify mentions of registration on ClinicalTrials.gov}
\usage{
.which_ct_2(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The study (EudraCT 2011‐001925‐26;
    ClinicalTrial.gov NCT01489592) was approved by the Ethics Committee of
    Rennes University Hospital."
}
