% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_ct_3}
\alias{.which_ct_3}
\title{Identify mentions of registration on ClinicalTrials.gov}
\usage{
.which_ct_3(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Registered on ClinicalTrials.gov (NCT12345678)."
}
