% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_disclosure_1}
\alias{.which_disclosure_1}
\title{Identify mentions of financial disclosures}
\usage{
.which_disclosure_1(article, dict)
}
\arguments{
\item{article}{The text as a vector of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract mentions of financial disclosures, e.g. "Financial disclosure:
    Nothing to disclose.
}
