% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_drks_1}
\alias{.which_drks_1}
\title{Identify mentions of registration on DRKS}
\usage{
.which_drks_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Identify mentions of registration on DRKS
}
