% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi_pmc.R
\name{.which_grants_1}
\alias{.which_grants_1}
\title{Identify receipt of grants}
\usage{
.which_grants_1(article)
}
\arguments{
\item{article}{The text as a vector of strings.}
}
\value{
Index of elements with phrase of interest
}
\description{
Identify mentions such as: "SS has received research grants from GSK."
}
