% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_not_registered_1}
\alias{.which_not_registered_1}
\title{Identify mentions of lack of registration}
\usage{
.which_not_registered_1(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "This trial and its protocol were not
    registered on a publicly accessible registry."
}
