% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_novelty.R
\name{.which_novelty_first_time_2}
\alias{.which_novelty_first_time_2}
\title{Identify "first time that" claims}
\usage{
.which_novelty_first_time_2(article)
}
\arguments{
\item{article}{A character vector of paragraphs.}
}
\value{
Integer index of matching elements.
}
\description{
Identify "first time that" claims
}
