% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_novelty.R
\name{.which_novelty_first_to_1}
\alias{.which_novelty_first_to_1}
\title{Identify "first to show/report/demonstrate" claims}
\usage{
.which_novelty_first_to_1(article)
}
\arguments{
\item{article}{A character vector of paragraphs.}
}
\value{
Integer index of matching elements.
}
\description{
Identify "first to show/report/demonstrate" claims
}
