% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_novelty.R
\name{.which_novelty_novel_1}
\alias{.which_novelty_novel_1}
\title{Identify "novel finding/approach" claims}
\usage{
.which_novelty_novel_1(article)
}
\arguments{
\item{article}{A character vector of paragraphs.}
}
\value{
Integer index of matching elements.
}
\description{
Identify "novel finding/approach" claims
}
