% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_protocol_1}
\alias{.which_protocol_1}
\title{Identify mentions of protocol}
\usage{
.which_protocol_1(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The complete study protocol has been
    published previously (Supplement 1)"
}
