% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_protocol_2}
\alias{.which_protocol_2}
\title{Identify mentions of protocol}
\usage{
.which_protocol_2(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Alliance for Clinical Trials in
    Oncology (formerly Cancer and Leukemia Group B) Protocol #369901"
}
