% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_reg_title_2}
\alias{.which_reg_title_2}
\title{Identify registration titles - specific with no negation}
\usage{
.which_reg_title_2(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Trial registration: ..."
}
