% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registered_1}
\alias{.which_registered_1}
\title{Identify generic mentions of registration}
\usage{
.which_registered_1(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "This study was approved by the local
    Scientific and Ethics Committees of IRCCS \"Saverio de Bellis\",
    Castellana Grotte (Ba), Italy, and it was part of a registered research
    on https://www.clinicaltrials.gov, reg. number: NCT01244945."
}
