% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registered_2}
\alias{.which_registered_2}
\title{Identify generic mentions of registration}
\usage{
.which_registered_2(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: " EPGP is registered with
    clinicaltrials.gov (NCT00552045)."
}
