% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registered_3}
\alias{.which_registered_3}
\title{Identify generic mentions of registration}
\usage{
.which_registered_3(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "registered with Clinical Trials
    (ChiCTR-IOR-14005438)"
}
