% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registered_4}
\alias{.which_registered_4}
\title{Identify generic mentions of registration}
\usage{
.which_registered_4(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "This registered study on
    www.clinicaltrials.gov (NCT01375270) was approved by the Ethics
    Committee of the Capital Region of Denmark (H-3-2010-127), and all
    subjects provided informed written consent to participate."
}
