% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registered_5}
\alias{.which_registered_5}
\title{Identify generic mentions of registration}
\usage{
.which_registered_5(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The Régression de l'Albuminurie dans
    la Néphropathie Drépanocytaire (RAND) study design was approved by the
    local ethics committee (Ref: DGRI CCTIRS MG/CP09.503, 9 July 2009) and
    registered at ClinicalTrials.gov (NCT01195818)."
}
