% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registration_4}
\alias{.which_registration_4}
\title{Identify generic mentions of registration}
\usage{
.which_registration_4(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The RAPiD trial's International
    Standard Randomised Controlled Trial Number (ISRCTN) registration is
    ISRCTN 49204710."
}
