% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register_pmc.R
\name{.which_registry_1}
\alias{.which_registry_1}
\title{Identify mentions of registry}
\usage{
.which_registry_1(article, dict)
}
\arguments{
\item{article}{A string or a list of strings.}

\item{dict}{A list of regular expressions for each concept.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Here, we describe a collaboration
    between an international group of patient organisations advocating for
    patients with atypical haemolytic uraemic syndrome (aHUS), the aHUS
    Alliance, and an international aHUS patient registry (ClinicalTrials.gov
    NCT01522183)."
}
