% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_replication.R
\name{.which_replication_replicat_1}
\alias{.which_replication_replicat_1}
\title{Identify replication claims using "replicate" terms}
\usage{
.which_replication_replicat_1(article)
}
\arguments{
\item{article}{A character vector of paragraphs.}
}
\value{
Integer index of matching elements.
}
\description{
Identify replication claims using "replicate" terms
}
