% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_replication.R
\name{.which_replication_validation_1}
\alias{.which_replication_validation_1}
\title{Identify "validation cohort/sample" claims}
\usage{
.which_replication_validation_1(article)
}
\arguments{
\item{article}{A character vector of paragraphs.}
}
\value{
Integer index of matching elements.
}
\description{
Identify "validation cohort/sample" claims
}
