% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{find_acknows}
\alias{find_acknows}
\title{Restrict to text after the body of the article and references}
\usage{
find_acknows(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the start and finish of this section.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
