% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{find_methods}
\alias{find_methods}
\title{Find the Methods section}
\usage{
find_methods(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Find the index of the start of the Methods section.
}
