% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{get_acknow_1}
\alias{get_acknow_1}
\title{Identify mentions of "Acknowledgement and"}
\usage{
get_acknow_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
