% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_anzctr_1}
\alias{get_anzctr_1}
\title{Identify mentions of registration on ANZCTR}
\usage{
get_anzctr_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The study was registered with ANZCTR
    (ACTRN12614001234567)."
}
