% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{get_common_1}
\alias{get_common_1}
\title{Get common phrases}
\usage{
get_common_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest.
}
