% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_common_3}
\alias{get_common_3}
\title{Get common phrases}
\usage{
get_common_3(article)

get_common_3(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Identify statements of the following type: "All authors are required to
    disclose all affiliations, funding sources and financial or management
    relationships that could be perceived as potential sources of bias. The
    authors disclosed none.

Identify statements of the following type: "All authors are required to
    disclose all affiliations, funding sources and financial or management
    relationships that could be perceived as potential sources of bias. The
    authors disclosed none.
}
