% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_common_5}
\alias{get_common_5}
\title{Get common phrases}
\usage{
get_common_5(article)

get_common_5(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Identify statements of the following type: "No specific sources of funding."

Identify statements of the following type: "No specific sources of funding."
}
