% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_developed_1}
\alias{get_developed_1}
\title{Identify mentions of developed}
\usage{
get_developed_1(article)

get_developed_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Return the index of statements such as: This publication was developed
    under Assistance Agreement No. 83563701-0 awarded by the U.S.
    Environmental Protection Agency to the University of Michigan.

Return the index of statements such as: This publication was developed
    under Assistance Agreement No. 83563701-0 awarded by the U.S.
    Environmental Protection Agency to the University of Michigan.
}
