% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_drks_1}
\alias{get_drks_1}
\title{Identify mentions of registration on DRKS}
\usage{
get_drks_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Registered at DRKS (DRKS00012345)."
}
