% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_fund_acknow_new}
\alias{get_fund_acknow_new}
\title{Identify mentions of funds in acknowledgements}
\usage{
get_fund_acknow_new(article)

get_fund_acknow_new(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.

Returns the index with the elements of interest. More generic than _1.
}
