% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_funded_ct_1}
\alias{get_funded_ct_1}
\title{Identify mentions of funding followed by NCT}
\usage{
get_funded_ct_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Funded by: the National Heart, Lung,
    and Blood Institute, the National Institute of Diabetes and Digestive and
    Kidney Disease, and others; SPECS ClinicalTrials.gov number, NCT00443599;
    Nutrition and Obesity Center at Harvard; NIH 5P30DK040561-17"
}
