% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_irct_1}
\alias{get_irct_1}
\title{Identify mentions of registration on IRCT (Iranian Registry of Clinical Trials)}
\usage{
get_irct_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Registered at IRCT (IRCT20120526009954N3)."
}
