% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_isrctn_1}
\alias{get_isrctn_1}
\title{Identify mentions of registration on ISRCTN}
\usage{
get_isrctn_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The study was registered at ISRCTN
    (ISRCTN12345678)."
}
