% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_not_registered_1}
\alias{get_not_registered_1}
\title{Identify mentions of lack of registration}
\usage{
get_not_registered_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "This trial and its protocol were not
    registered on a publicly accessible registry."
}
