% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{get_project_acknow}
\alias{get_project_acknow}
\title{Identify mentions of Grant numbers in the Funding/Acknowledgements}
\usage{
get_project_acknow(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
