% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_prospero_1}
\alias{get_prospero_1}
\title{Identify mentions of registration on PROSPERO}
\usage{
get_prospero_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "We registered the protocol for this
    meta-analysis with the PROSPERO database (www.crd.york.ac.uk/prosper
    o)—registration no. CRD42014015595."
}
