% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_protocol_1}
\alias{get_protocol_1}
\title{Identify mentions of protocol}
\usage{
get_protocol_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The complete study protocol has been
    published previously (Supplement 1)"
}
