% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{get_received_2}
\alias{get_received_2}
\title{Identify mentions of received}
\usage{
get_received_2(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
