% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_recipient_1}
\alias{get_recipient_1}
\title{Identify mentions of recepient}
\usage{
get_recipient_1(article)

get_recipient_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Returns the index of mentions such as: "Recipient of National Institutes of
    Health Grants AG044552, AI121621, and DK112365.

Returns the index of mentions such as: "Recipient of National Institutes of
    Health Grants AG044552, AI121621, and DK112365.
}
