% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_reg_title_4}
\alias{get_reg_title_4}
\title{Identify registration titles - specific}
\usage{
get_reg_title_4(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Clinical trial registration detail:"
}
