% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registered_1}
\alias{get_registered_1}
\title{Identify generic mentions of registration}
\usage{
get_registered_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "This study was approved by the local
    Scientific and Ethics Committees of IRCCS \"Saverio de Bellis\",
    Castellana Grotte (Ba), Italy, and it was part of a registered research
    on https://www.clinicaltrials.gov, reg. number: NCT01244945."
}
