% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registered_2}
\alias{get_registered_2}
\title{Identify generic mentions of registration}
\usage{
get_registered_2(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: " EPGP is registered with
    clinicaltrials.gov (NCT00552045)."
}
