% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registered_4}
\alias{get_registered_4}
\title{Identify generic mentions of registration}
\usage{
get_registered_4(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "This registered study on
    www.clinicaltrials.gov (NCT01375270) was approved by the Ethics
    Committee of the Capital Region of Denmark (H-3-2010-127), and all
    subjects provided informed written consent to participate."
}
