% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registered_5}
\alias{get_registered_5}
\title{Identify generic mentions of registration}
\usage{
get_registered_5(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "The Régression de l'Albuminurie dans
    la Néphropathie Drépanocytaire (RAND) study design was approved by the
    local ethics committee (Ref: DGRI CCTIRS MG/CP09.503, 9 July 2009) and
    registered at ClinicalTrials.gov (NCT01195818)."
}
