% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registration_1}
\alias{get_registration_1}
\title{Identify generic mentions of registration}
\usage{
get_registration_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Trial registration number is
    TCTR20151021001."
}
