% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registration_3}
\alias{get_registration_3}
\title{Identify generic mentions of registration}
\usage{
get_registration_3(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Public clinical trial registration
    (www.clinicaltrials.gov) ID# NCT02720653"
}
