% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{get_registry_1}
\alias{get_registry_1}
\title{Identify mentions of registry}
\usage{
get_registry_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Extract the index of mentions such as: "Here, we describe a collaboration
    between an international group of patient organisations advocating for
    patients with atypical haemolytic uraemic syndrome (aHUS), the aHUS
    Alliance, and an international aHUS patient registry (ClinicalTrials.gov
    NCT01522183)."
}
