% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_support_10}
\alias{get_support_10}
\title{Identify mentions of support}
\usage{
get_support_10(article)

get_support_10(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Return the index of support statements such as: We are grateful to the
    National Institute of Mental Health (MH091070: PI’s Dante Cicchetti and
    Sheree L. Toth) for support of this work.

Return the index of support statements such as: We are grateful to the
    National Institute of Mental Health (MH091070: PI’s Dante Cicchetti and
    Sheree L. Toth) for support of this work.
}
