% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{get_support_3}
\alias{get_support_3}
\title{Identify mentions of support}
\usage{
get_support_3(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
