% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_support_6}
\alias{get_support_6}
\title{Identify mentions of support}
\usage{
get_support_6(article)

get_support_6(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Return the index of support statements such as: We gratefully acknowledge
    support from the UK Engineering and Physical Sciences Research Council.

Return the index of support statements such as: We gratefully acknowledge
    support from the UK Engineering and Physical Sciences Research Council.
}
