% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_support_7}
\alias{get_support_7}
\title{Identify mentions of support}
\usage{
get_support_7(article)

get_support_7(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Return the index of support statements such as: Support was provided by the
    U.S. Department of Agriculture (USDA) Forest Service.

Return the index of support statements such as: Support was provided by the
    U.S. Department of Agriculture (USDA) Forest Service.
}
