% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_support_9}
\alias{get_support_9}
\title{Identify mentions of support}
\usage{
get_support_9(article)

get_support_9(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Return the index of support statements such as: The US Environmental
    Protection Agency’s (USEPA) Office of Research and Development funded and
    managed the research described in the present study.

Return the index of support statements such as: The US Environmental
    Protection Agency’s (USEPA) Office of Research and Development funded and
    managed the research described in the present study.
}
