% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_thank_2}
\alias{get_thank_2}
\title{Identify mentions of thank you statements}
\usage{
get_thank_2(article)

get_thank_2(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Returns the index with the elements related to "The authors acknowledge
    Worldwide Cancer Research (AICR) 15-1002; Blueprint 282510; ..."

Returns the index with the elements related to "The authors acknowledge
    Worldwide Cancer Research (AICR) 15-1002; Blueprint 282510; ..."
}
