% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{is_funding}
\alias{is_funding}
\title{Identify funding statements}
\usage{
is_funding(filename)
}
\arguments{
\item{filename}{A List with paragraphs of interest.}
}
\value{
A dataframe indicating whether a funding statement has been
    identified and the funding statement.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
