% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{negate_absence_1}
\alias{negate_absence_1}
\title{Avoid mentions that no funding information was provided}
\usage{
negate_absence_1(article)

negate_absence_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
