% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{negate_conflict_1}
\alias{negate_conflict_1}
\title{Avoid financial that is part of COI statements}
\usage{
negate_conflict_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.
}
\description{
Returns the index with the elements of interest. More generic than _1.
}
