% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{negate_reg_title_1}
\alias{negate_reg_title_1}
\title{Negate unwanted mentions of title}
\usage{
negate_reg_title_1(article)
}
\arguments{
\item{article}{A string or a list of strings.}
}
\value{
Index of element with phrase of interest
}
\description{
Negate title mentions that refer to unwanted text.
}
