% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{obliterate_comma_1}
\alias{obliterate_comma_1}
\title{Remove commas}
\usage{
obliterate_comma_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Removes commas to simplify regular expressions.
}
