% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{obliterate_conflict_1}
\alias{obliterate_conflict_1}
\title{Remove mentions of COIs that may cause FPs}
\usage{
obliterate_conflict_1(articles)
}
\arguments{
\item{articles}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Returns the text without potentially misleading mentions of COIs.
}
