% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund_pmc.R
\name{obliterate_conflict_2}
\alias{obliterate_conflict_2}
\title{Remove COI statements with language relevant to funding statements}
\usage{
obliterate_conflict_2(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Returns the text without mentions of COIs related to industry ties.
}
