% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund_pmc.R
\name{obliterate_misleading_fund_1}
\alias{obliterate_misleading_fund_1}
\title{Remove mentions of commonly misleading funding statements}
\usage{
obliterate_misleading_fund_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Returns the text without potentially misleading phrases related to funding.
}
