% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{obliterate_refs_1}
\alias{obliterate_refs_1}
\title{Remove references}
\usage{
obliterate_refs_1(articles)
}
\arguments{
\item{articles}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without misleading fullstops.
}
\description{
Returns the list of paragraphs without references.
}
