% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{obliterate_semicolon_1}
\alias{obliterate_semicolon_1}
\title{Remove semicolons when within parentheses}
\usage{
obliterate_semicolon_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.
}
\description{
Removes mentions such as: "guidelines for diagnostic studies (trial
    registered at www.clinicaltrial.gov; NCT01697930)."
}
