% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_all.R
\name{rt_all}
\alias{rt_all}
\title{Identify and extract statements of COI, Funding and Registration.}
\usage{
rt_all(filename)
}
\arguments{
\item{filename}{The name of the TXT as a string.}
}
\value{
A dataframe of results.
}
\description{
Takes a TXT file and examines whether any statements of Conflicts of Interest
    (COI), Funding or Protocol Registration exist. If any such statements are
    found, it also extracts the relevant text.
}
\examples{
\dontrun{
# Path to TXT.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract indicators of transparency.
results_table <- rt_all(filepath)
}
}
