% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_data_code.R
\name{rt_data_code}
\alias{rt_data_code}
\title{Identify and extract Data and Code statements in TXT files.}
\usage{
rt_data_code(filename)
}
\arguments{
\item{filename}{The name of the TXT file as a string.}
}
\value{
A dataframe of results.
}
\description{
Takes a TXT file and returns data related to the presence of Data and/or Code
    statements, including whether Data and/or Code statements exist. If such
    statements exist, it extracts them.
}
\examples{
\dontrun{
# Path to PMC XML.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract meta-data and indicators of transparency.
results_table <- rt_data(filepath)
}
}
