% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_data_code_pmc.R
\name{rt_data_code_pmc_list}
\alias{rt_data_code_pmc_list}
\title{Identify and extract Data and Code sharing from a list of PMC XML files.}
\usage{
rt_data_code_pmc_list(filenames, remove_ns = T, specificity = "low")
}
\arguments{
\item{filenames}{A list of the PMC XML filenames as strings.}

\item{remove_ns}{TRUE if an XML namespace exists, else FALSE (default).}

\item{specificity}{Retained for backward compatibility; see
\code{\link{rt_data_code_pmc}}.}
}
\value{
A dataframe of results, one row per file.
}
\description{
Takes a list of PMC XML files and returns data related to the presence of
    Data or Code, including whether Data or Code have been shared. If Data
    or Code exist, it will extract the relevant text for each.
}
\examples{
\dontrun{
# Paths to PMC XML files
filepath <- system.file(
  "extdata", "PMID32171256-PMC7071725.xml", package = "rtransparent"
)
filepaths <- list(filepath)

# Identify and extract indicators of data and code sharing
results_table <- rt_data_code_pmc_list(filepaths, remove_ns = TRUE)
}
}
