% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_read_pdf.R
\name{rt_read_pdf}
\alias{rt_read_pdf}
\title{Covert PDF files into TXT files.}
\usage{
rt_read_pdf(filepath)
}
\arguments{
\item{filepath}{The path to the TXT file as a string.}
}
\value{
A character object.
}
\description{
Takes a path to a PDF file and returns a TXT file. Note that there are a
    number of such converters and different converters return TXT files
    formatted differently. The functions within this package were created
    to work well with the converter used within this function (poppler).
    This function was taken from the package `oddpub` by Nico Riedel and
    hereby modified for the purposes of this package.
}
\examples{
\dontrun{
# Path to PDF file
filepath <- "../inst/extdata/PMID30457984-PMC6245499.txt"

# Convert into string
article_txt <- rt_read_pdf(filepath)
}
}
