% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{rt_register}
\alias{rt_register}
\title{Identify and extract Registration statements in TXT files.}
\usage{
rt_register(filename)
}
\arguments{
\item{filename}{The name of the TXT file as a string.}
}
\value{
A dataframe of results.
}
\description{
Takes a TXT file and returns data related to the presence of a Registration
    statement, including whether a Registration statement exists. If a
    Registration statement exists, it extracts it.
}
\examples{
\dontrun{
# Path to PMC XML.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract meta-data and indicators of transparency.
results_table <- rt_register(filepath)
}
}
