% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_replication.R
\name{rt_replication}
\alias{rt_replication}
\title{Identify whether a study includes a replication component in TXT files.}
\usage{
rt_replication(filename)
}
\arguments{
\item{filename}{The name of the TXT file as a string.}
}
\value{
A tibble of results. It returns the filename, PMID (if it was part
    of the file name), whether a replication component was found, the text
    identified, and whether each pattern-matching function identified
    relevant text or not.
}
\description{
Takes a TXT file and returns data related to the presence of a replication
    or validation component, including whether such a component exists.
    Replication is defined as the study independently confirming findings
    from a prior study in a new sample.
}
\examples{
\dontrun{
# Path to TXT file.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract replication components.
results_table <- rt_replication(filepath)
}
}
