% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_replication_pmc.R
\name{rt_replication_pmc}
\alias{rt_replication_pmc}
\title{Identify and extract replication components in PMC XML files.}
\usage{
rt_replication_pmc(filename, remove_ns = FALSE)
}
\arguments{
\item{filename}{The name of the PMC XML as a string.}

\item{remove_ns}{TRUE if an XML namespace exists, else FALSE (default).}
}
\value{
A tibble of results. It returns the unique identifiers of the
    article, whether a replication component was found, the relevant text
    and whether each pattern-matching function identified relevant text.
}
\description{
Takes a PMC XML file and returns data related to the presence of a
    replication or validation component, including whether such a component
    exists and the relevant text. Replication is defined as the study
    independently confirming findings from a prior study in a new sample.
}
\examples{
\dontrun{
# Path to PMC XML.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.xml"

# Identify and extract replication components.
results_table <- rt_replication_pmc(filepath, remove_ns = TRUE)
}
}
