% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtransparent.R
\docType{package}
\name{rtransparent-package}
\alias{rtransparent}
\alias{rtransparent-package}
\title{rtransparent: Identify indicators of transparency in the published literature}
\description{
It converts a PDF to a text file and then checks for the presence of certain
indicators of transparency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/choxos/rtransparent}
  \item \url{https://choxos.github.io/rtransparent}
  \item Report bugs at \url{https://github.com/choxos/rtransparent/issues}
}

}
\author{
\strong{Maintainer}: Ahmad Sofi-Mahmudi \email{a.sofimahmudi@gmail.com} (\href{https://orcid.org/0000-0001-6829-0823}{ORCID})

Authors:
\itemize{
  \item Ahmad Sofi-Mahmudi \email{a.sofimahmudi@gmail.com} (\href{https://orcid.org/0000-0001-6829-0823}{ORCID})
  \item Stylianos Serghiou \email{stelios.serghiou@gmail.com}
}

}
