% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{.max_words}
\alias{.max_words}
\title{Place boundaries around words}
\usage{
.max_words(x, n_max = 3)
}
\arguments{
\item{x}{A vector of strings.}

\item{n_max}{Number of maximum words allowed}
}
\value{
A vector of bounded strings.
}
\description{
Returns each string with boundaries around it.
}
