% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_project_acknow}
\alias{get_project_acknow}
\title{Identify mentions of project numbers.}
\usage{
get_project_acknow(article)

get_project_acknow(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Returns the index of mentions such as: "NIH Project no. AB123."

Returns the index of mentions such as: "Remerciements Cette étude a été
    financée par l’Institut de Recherches Médicales et d’Etudes des Plantes
    Médicinales (IMPM)."
}
