% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{get_support_1}
\alias{get_support_1}
\title{Identify mentions of support}
\usage{
get_support_1(article)

get_support_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The index of the paragraph of interest.

The index of the paragraph of interest.
}
\description{
Identifies mentions of "This work was funded by ..." and of "This work was
    completed using funds from ..."

Identifies mentions of "This work was funded by ..." and of "This work was
    completed using funds from ..."
}
