% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_fund.R, R/rt_fund_pmc.R
\name{obliterate_conflict_1}
\alias{obliterate_conflict_1}
\title{Remove mentions of COIs that may cause FPs}
\usage{
obliterate_conflict_1(article)

obliterate_conflict_1(article)
}
\arguments{
\item{article}{A List with paragraphs of interest.}
}
\value{
The list of paragraphs without mentions of financial COIs.

The list of paragraphs without mentions of financial COIs.
}
\description{
Returns the text without potentially misleading mentions of COIs.

Returns the text without potentially misleading mentions of COIs.
}
