% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_all.R
\name{rt_all}
\alias{rt_all}
\title{Identify and extract statements of COI, Funding and Registration.}
\usage{
rt_all(filename)
}
\arguments{
\item{filename}{The name of the TXT as a string.}
}
\value{
A dataframe of results. It returns the PMID of the article (if this
    was part of the filename and preceded by "PMID"), whether each indicator
    of COI, Funding or Registration was identified, the identified text and
    whether each labelling function identified relevant text or not. The
    labeling functions are returned to add flexibility in how this package is
    used; for example, future definitions of Registration may differ from the
    one we used. If a labelling function returns NA it means that it was not
    run.
}
\description{
Takes a TXT file and examines whether any statements of Conflicts of Interest
    (COI), Funding or Protocol Registration exist. If any such statements are
    found, it also extracts the relevant text.
}
\examples{
\dontrun{
# Path to TXT.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract indicators of transparency.
results_table <- rt_all(filepath)
}
}
