% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_coi.R
\name{rt_coi}
\alias{rt_coi}
\title{Identify and extract Conflicts of Interest (COI) statements in TXT files.}
\usage{
rt_coi(filename)
}
\arguments{
\item{filename}{The name of the TXT file as a string.}
}
\value{
A dataframe of results. It returns the filename, PMID (if it was part
    of the file name), whether a COI was found and the text identified.
}
\description{
Takes a TXT file and returns data related to the presence of a COI
    statement, including whether a COI statement exists. If a COI statement
    exists, it extracts it.
}
\examples{
\dontrun{
# Path to PMC XML.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract meta-data and indicators of transparency.
results_table <- rt_coi(filepath)
}
}
