% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_register.R
\name{rt_register}
\alias{rt_register}
\title{Identify and extract Registration statements in TXT files.}
\usage{
rt_register(filename)
}
\arguments{
\item{filename}{The name of the TXT file as a string.}
}
\value{
A dataframe of results. It returns the PMID (if this was part of the
    filename and preceded by PMID), whether a registration statement was
    found, the identified statement, whether the text was deemed relevant
    (e.g. contained the word registration), whether a Methods section was
    identified, whether an NCT number was identified, whether a registration
    was explicitly identified (defunct) and whether each labeling function
    identified a relevant text or not. The labeling functions are returned to
    add flexibility in how this package is used; for example, future
    definitions of Registration may differ from the one we used.
}
\description{
Takes a TXT file and returns data related to the presence of a Registration
    statement, including whether a Registration statement exists. If a
    Registration statement exists, it extracts it.
}
\examples{
\dontrun{
# Path to PMC XML.
filepath <- "../inst/extdata/00003-PMID26637448-PMC4737611.txt"

# Identify and extract meta-data and indicators of transparency.
results_table <- rt_register(filepath)
}
}
