% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtransparent.R
\docType{package}
\name{rtransparent-package}
\alias{rtransparent}
\alias{rtransparent-package}
\title{rtransparent: Identify indicators of transparency in the published literatrue}
\description{
It converts a PDF to a text file and then checks for the presence of certain
indicators of transparency.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/serghiou/rtransparent}
  \item Report bugs at \url{http://github.com/serghiou/rtransparent/issues}
}

}
\author{
\strong{Maintainer}: Stylianos Serghiou \email{stelios.serghiou@gmail.com}

}
